requirements-gpu.txt
aniso8601==1.2.1
beautifulsoup4==4.6.0
bs4==0.0.1
click==6.7
Flask==0.12.2
Flask-Mobility==0.1.1
Flask-RESTful==0.3.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
numpy==1.13.0
prettytable==0.7.2
protobuf==3.3.0
psycopg2==2.7.1
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0
tensorflow-gpu==1.0.0
Werkzeug==0.12.2
